requirements_tests.txt
astor==0.8.1
autogen_agentchat==0.4.8
autogen_core==0.4.8
autogen_ext==0.4.8
composio_autogen==0.7.5
composio_crewai==0.7.5
composio_langchain==0.7.5
composio_llamaindex==0.7.5
llama-index-tools-duckduckgo==0.3.0
llama-index-tools-wikipedia==0.3.0
crewai_tools==0.36.0
duckduckgo_search==6.4.2
Faker==36.1.1
Jinja2==3.1.5
langchain_community==0.3.19
langchain_core==0.3.41
langchain_experimental==0.3.4
langchain_google_community==2.0.7
langgraph==0.3.5
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
parameterized==0.9.0
pydantic==2.10.6
pydantic_yaml==1.4.0
PyGithub==2.6.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
seaborn==0.13.2
typing_extensions==4.12.2
json_schema_to_pydantic